Explain coaching styles in the profile and default new notebooks to Quick.
Students can see what Quick and Strict mean before choosing; the selected block highlights like Appearance. New notebooks still store Quick as short and Strict as long.

+++ backend/learning/journey.py
@@ -16,7 +16,7 @@ from backend.specialists.review_orchestration import (
 )
 
 RESPONSE_DETAILS = ("short", "long")
-DEFAULT_RESPONSE_DETAIL = "long"
+DEFAULT_RESPONSE_DETAIL = "short"
 _STAGE_DECISION = re.compile(
     r"<!--\s*stage\s*:\s*(advance|stay)\s*-->",
     re.IGNORECASE,
